pyproject.toml
[tool.poetry]
name = "slide-tc"
version = "0.1.0"
description = "solve slide Tcaptcha"
authors = ["aioqzone <zzzzss990315@gmail.com>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^10.0.1"
numpy = "^1.22.3"

# dependency groups
[tool.poetry.group.test]
optional = false

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "~0.21.0"
aioqzone = "^1.5.6.dev1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
black = "*"
isort = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
